restructured state sync pools
